ENH: 'setup-user-bin.sh' skips lines that start with hash mark

+++ ee-ik1614lx/setup/setup-user-bin.sh
@@ -34,15 +34,26 @@ DEFAULT_BIN_SCRIPTS="${EE_IK1614LX_DEFAULTS_HOME}/default_user_bin_scripts.txt"
 
 echo "`INFO $_FILE_NAME` Configuring convenience scripts in [${USER_BIN_HOME}]."
 
-# '|| [[ -n $line ]]' prevents the last line from being ignored if it doesn't end with a \n
+# '|| [[ -n $line ]]' prevents the last line being ignored if it doesn't end with a \n
 while read -r line || [[ -n "$line" ]]; do
-
-    script_file_name="$line"
-    printf "\t Copying `green $script_file_name` \n"
-    cp $CSP_ICT_HOME/scripts/${script_file_name} ${USER_BIN_HOME}/${script_file_name}
 
     # Skip empty lines if they exist
     [ -z "$line" ] && continue
+
+    # Skip lines that start with a hash mark
+    [[ ${line} == \#* ]] && echo "Skip copying `yellow $line`" && continue
+
+    script_file_name="$line"
+    target=$CSP_ICT_HOME/scripts/${script_file_name}
+    destination=${USER_BIN_HOME}/${script_file_name}
+
+    if [ ! -f $target ]; then
+        echo "`yellow WARNING:` Couldn't copy [`red $script_file_name`], it doesn't exist."
+    else
+        echo "Copying `green $script_file_name`"
+        cp $CSP_ICT_HOME/scripts/${script_file_name} ${USER_BIN_HOME}/${script_file_name}
+    fi
+
 
 done < "$DEFAULT_BIN_SCRIPTS"
